shop/requirements.txt
asgiref==3.2.7
boto3==1.13.1
botocore==1.16.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
defusedxml==0.6.0
Django==3.0.5
django-countries==6.1.2
django-material==1.6.3
django-money==1.1
django-sendgrid-v5==0.8.1
django-storages==1.9.1
docutils==0.15.2
ecdsa==0.15
future==0.18.2
gunicorn==20.0.4
idna==2.9
jmespath==0.9.5
money==1.3.0
oauthlib==3.1.0
Pillow==7.1.1
psycopg2-binary==2.8.5
py-moneyed==0.8.0
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-http-client==3.2.7
python-jose==3.1.0
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
sendgrid==6.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
stripe==2.45.0
urllib3==1.25.8
whitenoise==5.0.1
XlsxWriter==1.2.8

pytest~=5.4.2
pip~=19.0.3